Refactor constants and calculations to use option-lab namespace
- Moved calculations and constants from `mint` to `option-lab` directory
- Updated import paths across multiple components to reflect new directory structure
- Removed unused imports and simplified component dependencies
- Maintained existing calculation and constant logic during refactoring
